chore(docs): move syslogng config to separate file
