+++ docs/configuration/crds/v1beta1/_index.md
@@ -26,8 +26,9 @@ For more information please click on the name
 | **[OutputSpec](output_types/)** | OutputSpec defines the desired state of Output | v1beta1 |
 | **[SyslogNGClusterFlow](syslogng_clusterflow_types/)** | SyslogNGClusterFlow is the Schema for the syslog-ng clusterflows API | v1beta1 |
 | **[SyslogNGClusterOutput](syslogng_clusteroutput_types/)** | SyslogNGClusterOutput is the Schema for the syslog-ng clusteroutputs API | v1beta1 |
+| **[SyslogNG](syslogng_config_types/)** | SyslogNGConfig is a standalone reference to the desired SyslogNG configuration | v1beta1 |
 | **[SyslogNGFlowSpec](syslogng_flow_types/)** | SyslogNGFlowSpec is the Kubernetes spec for SyslogNGFlows | v1beta1 |
 | **[SyslogNGOutputSpec](syslogng_output_types/)** | SyslogNGOutputSpec defines the desired state of SyslogNGOutput | v1beta1 |
-| **[SyslogNG](syslogng_types/)** | SyslogNG is a reference to the desired SyslogNG state | v1beta1 |
+| **[SyslogNGSpec](syslogng_types/)** | SyslogNGSpec defines the desired state of SyslogNG | v1beta1 |
 </center>
